Fix Debug Problem

+++ service/src/main/java/org/openo/sdno/vpc/nbi/SubnetNbiServiceImpl.java
@@ -20,11 +20,11 @@ import javax.annotation.Resource;
 
 import org.openo.baseservice.remoteservice.exception.ServiceException;
 import org.openo.sdno.overlayvpn.model.common.enums.ActionStatus;
+import org.openo.sdno.overlayvpn.result.ResultRsp;
 import org.openo.sdno.overlayvpn.model.netmodel.vpc.Subnet;
 import org.openo.sdno.overlayvpn.model.netmodel.vpc.Vpc;
 import org.openo.sdno.vpc.nbi.inf.ISubnetNbiService;
 import org.openo.sdno.vpc.nbi.inf.IVpcNbiService;
-import org.openo.sdno.vpc.sbi.SubnetSbiServiceImpl;
 import org.openo.sdno.vpc.sbi.inf.ISubnetSbiService;
 import org.openo.sdno.vpc.util.DaoUtils;
 import org.slf4j.Logger;
@@ -33,94 +33,74 @@ import org.slf4j.LoggerFactory;
 /**
  * VPC service NBI implementation class.
  * <br>
- * <p>
- * </p>
  *
  * @author
  * @version SDNO 0.5 2016-7-07
  */
-// TODO(mrkanag) kindly add interface :)
 public class SubnetNbiServiceImpl implements ISubnetNbiService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(SubnetNbiServiceImpl.class);
 
     @Resource
-    private ISubnetSbiService service = new SubnetSbiServiceImpl();
+    private ISubnetSbiService service;
 
     @Resource
-    private IVpcNbiService vpcService = new VpcNbiServiceImpl();
-
-    public ISubnetSbiService getService() {
-        return this.service;
-    }
+    private IVpcNbiService vpcService;
 
     public void setService(ISubnetSbiService service) {
         this.service = service;
-    }
-
-    public IVpcNbiService getVpcService() {
-        return this.vpcService;
     }
 
     public void setVpcService(IVpcNbiService vpcService) {
         this.vpcService = vpcService;
     }
 
-    /**
-     * Creates the subnet
-     * <br>
-     *
-     * @param subnet Subnet
-     * @return
-     * @throws ServiceException
-     * @since SDNO 0.5
-     */
     @Override
     public Subnet create(Subnet subnet) throws ServiceException {
-        LOGGER.debug("START");
 
-        // TODO(mrkanag) make sure subnet.getVpcId() already exist in db
+        // Insert DB
         subnet.setStatus(ActionStatus.CREATING.getName());
-        Subnet subnetLocal = DaoUtils.insert(subnet);
+        DaoUtils.insert(subnet);
 
-        Vpc vpc = this.vpcService.get(subnetLocal.getVpcId());
-        // TODO(mrkanag) set the createdAt time.
+        // Get Vpc from DB
+        Vpc vpc = this.vpcService.get(subnet.getVpcId());
 
+        Subnet subnetRsp = null;
         try {
-            // TODO(mrkanag) make this async
-            subnetLocal = this.service.create(vpc.getOsControllerId(), subnetLocal);
+            subnetRsp = this.service.create(vpc.getOsControllerId(), subnet);
         } catch(ServiceException e) {
-            LOGGER.error("Failed to create subnet " + subnetLocal.getCidr());
-            subnetLocal.setStatus(ActionStatus.CREATE_EXCEPTION.getName());
+            LOGGER.error("Failed to create subnet " + subnet.getCidr());
+            subnet.setStatus(ActionStatus.CREATE_EXCEPTION.getName());
             DaoUtils.update(subnet, "status");
             throw e;
         }
 
-        subnetLocal.setStatus(ActionStatus.NORMAL.getName());
-        subnetLocal = DaoUtils.update(subnetLocal, "status");
+        subnetRsp.setStatus(ActionStatus.NORMAL.getName());
+        DaoUtils.update(subnetRsp, "status");
 
-        LOGGER.debug("END " + subnetLocal.getUuid() + " is created successfully");
-
-        return subnetLocal;
+        LOGGER.info("END " + subnetRsp.getUuid() + " is created successfully");
+        return subnetRsp;
     }
 
-    /**
-     * Deletes subnet
-     * <br>
-     *
-     * @param subnetId Subnet Id
-     * @throws ServiceException
-     * @since SDNO 0.5
-     */
     @Override
     public void delete(String subnetId) throws ServiceException {
         LOGGER.debug("START");
 
         Subnet subnet = this.get(subnetId);
+        if(null == subnet) {
+            LOGGER.info("Subnet not exist.");
+            return;
+        }
+
         Vpc vpc = this.vpcService.get(subnet.getVpcId());
+        if(null == vpc) {
+            LOGGER.info("Vpc not exist.");
+            return;
+        }
 
         subnet.setStatus(ActionStatus.DELETING.getName());
-        subnet = DaoUtils.insert(subnet);
+        DaoUtils.update(subnet, "status");
+
         try {
             // TODO(mrkanag) make this async
             this.service.delete(vpc.getOsControllerId(), subnetId);
@@ -131,21 +111,19 @@ public class SubnetNbiServiceImpl implements ISubnetNbiService {
             throw e;
         }
 
-        // TODO(mrkanag) handle error
+        DaoUtils.delete(Subnet.class, subnetId);
+
         LOGGER.debug("END " + subnetId + " is deleted successfully");
     }
 
-    /**
-     * Retrieves subnet.
-     * <br>
-     *
-     * @param subnetId Subnet id.
-     * @return
-     * @throws ServiceException
-     * @since SDNO 0.5
-     */
     @Override
     public Subnet get(String subnetId) throws ServiceException {
-        return DaoUtils.get(Subnet.class, subnetId);
+
+        ResultRsp<Subnet> subnetRsp = DaoUtils.get(Subnet.class, subnetId);
+        if(!subnetRsp.isSuccess()) {
+            LOGGER.error("Failed to query subnet " + subnetId);
+            throw new ServiceException("Failed to query subnet " + subnetId);
+        }
+        return subnetRsp.getData();
     }
 }
